Backend: Adding Article service and controller and assuring that user its author and adding role author to the enumRole

+++ Adoptek-BackEnd/AdoptekSpring/src/main/java/com/example/adoptekspring/service/implementation/ArticleServiceImpl.java
@@ -3,26 +3,34 @@ package com.example.adoptekspring.service.implementation;
 import com.example.adoptekspring.domain.Article;
 import com.example.adoptekspring.repository.ArticleRepository;
 import com.example.adoptekspring.service.ArticleService;
+import com.example.adoptekspring.service.AuthorityService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class ArticleServiceImpl implements ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final AuthorityService authorityService;
 
-    public ArticleServiceImpl(ArticleRepository articleRepository) {
-        this.articleRepository = articleRepository;
-    }
 
     @Override
     public Article createArticle(Article article) {
+        if (!authorityService.hasRole("Author")) {
+            throw new RuntimeException("You are not authorized to create an article");
+        }
+
         return articleRepository.save(article);
     }
 
     @Override
     public Article updateArticle(Article article) {
+        if (!authorityService.hasRole("Author")) {
+            throw new RuntimeException("You are not authorized to create an article");
+        }
         Article article1 = articleRepository.findById(article.getId()).orElseThrow(() -> new RuntimeException("Article not found"));
         article1.setTitle(article.getTitle());
         article1.setContent(article.getContent());
@@ -32,6 +40,9 @@ public class ArticleServiceImpl implements ArticleService {
 
     @Override
     public void deleteArticle(Long id) {
+        if (!authorityService.hasRole("Author")) {
+            throw new RuntimeException("You are not authorized to create an article");
+        }
         articleRepository.deleteById(id);
     }
